Fix base plant logic issues and unit tests
The Base Plant logic was originally based upon timestamps were
initialized at -1 to indicate no data had been updated. This would cause
issues when combined with ROS so we changed the initialization value of
timestamps to 0. This commit fixes the resulting logic bugs that relied
on timestamps of -1 to determine whether or not to run various code.
Basic things include no calls to pubControl() until there a control
sequence has been calculated, and waiting for a new state logic.

- Change explicit calls to exit() into throwing exceptions
- Existing unit tests have been updated
- New unit tests written to explicitly test the new logic adjustments.